Création de compte, infos de base fonctionnelles

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -25,7 +25,7 @@ class CreateUsersTable extends Migration
             // Ajout des nouvelles colonnes:
             $table->date('birth_date')->useCurrent();
             $table->integer('max_distance');
-            $table->float('hourly_rate', 3, 2);
+            $table->integer('hourly_rate');
         });
     }
